% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gaussian_gtwid}
\alias{gaussian_gtwid}
\title{Fourier transform of Gaussian density}
\usage{
gaussian_gtwid(sd)
}
\arguments{
\item{sd}{Standard deviation}
}
\value{
Object of class \code{spedecon_gtwid}
}
\description{
Returns a \code{spedecon_gtwid} object representing the Fourier transform of a mean-zero Gaussian density
}
\examples{
gtwid <- gaussian_gtwid(sd = 1)
}
