% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{laplace_gtwid}
\alias{laplace_gtwid}
\title{Fourier transform of Laplace density}
\usage{
laplace_gtwid(b)
}
\arguments{
\item{b}{Scale parameter}
}
\value{
Object of class \code{spedecon_gtwid}
}
\description{
Returns a \code{spedecon_gtwid} object representing the Fourier transform of a mean-zero Laplace density with scale \code{b}
}
\examples{
gtwid <- laplace_gtwid(b = 1)
}
