% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_spedecon_gtwid}
\alias{new_spedecon_gtwid}
\title{Creates object of class \code{spedecon_gtwid}}
\usage{
new_spedecon_gtwid(gtwid, family)
}
\arguments{
\item{gtwid}{Function representing the Fourier transform}

\item{family}{List with at least one entry \code{family[["family"]]} naming the family of distributions, and possibly other entries stating the values of the parameters in that family.}
}
\value{
Object of class \code{spedecon_gtwid}
}
\description{
Constructor for class \code{spedecon_gtwid}. Use helper functions \code{\link[=gaussian_gtwid]{gaussian_gtwid()}}, \code{\link[=laplace_gtwid]{laplace_gtwid()}}, and \code{\link[=uniform_gtwid]{uniform_gtwid()}} instead whenever possible.
}
\details{
The \code{spedecon_gtwid} class is meant to represent the Fourier transform of a probability density.
The basic type is a function.
It also has a \code{family} attribute which can hold the name and parameters of the family of distributions.
}
\seealso{
Use \code{\link[=gaussian_gtwid]{gaussian_gtwid()}}, \code{\link[=laplace_gtwid]{laplace_gtwid()}}, or \code{\link[=uniform_gtwid]{uniform_gtwid()}} instead whenever possible.
}
