% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_spedecon_spline_sped_fit}
\alias{new_spedecon_spline_sped_fit}
\title{Creates object of class \code{spedecon_spline_sped_fit}}
\usage{
new_spedecon_spline_sped_fit(coef, basis, alpha, constraint)
}
\arguments{
\item{coef}{Numeric vector; the coefficients of the spline.}

\item{basis}{Object of class \code{spedecon_spline_basis}, representing a basis for the spline space.}

\item{alpha}{Positive numeric, the alpha that was used for the fit.}

\item{constraint}{The type of constraint used.}
}
\value{
Object of class \code{spedecon_spline_sped_fit}
}
\description{
Internal use only. Constructor for class \code{spedecon_spline_sped_fit}
}
\details{
The basic type of an object of type \code{spedecon_spline_sped_fit} is a function; one can therefore evaluate, plot, etc. and ignore the other attributes if desired.
The function is represented as a spline, and has attributes \code{coef} and \code{basis}, which represents the coefficients and basis respectively.
\code{coef} is a numeric vector, while \code{basis} is an object of class \code{spedecon_spline_basis}, which is essentially just a list holding the knots and order of the spline space.
A \code{spedecon_spline_sped_fit} object also has attributes \code{alpha} and \code{constraint} which record the penalty parameter and constraint method used for the fit.
}
