% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniform_gtwid}
\alias{uniform_gtwid}
\title{Fourier transform of Uniform density}
\usage{
uniform_gtwid(a)
}
\arguments{
\item{a}{Half-width}
}
\value{
Object of class \code{spedecon_gtwid}
}
\description{
Returns a \code{spedecon_gtwid} object representing the Fourier transform of a Uniform[-a,a] density
}
\examples{
gtwid <- uniform_gtwid(a = 1)
}
